+++ utils/data_storage.py
@@ -1,0 +1,4 @@
+# utils/data_storage.py
+
+
+user_booking_data = {}
